I renamed files and I added a log-out button.
