+++ src/main/java/app/GameApplication.java
@@ -3,7 +3,7 @@ package app;
 import data_access.DBSearchDataAccessObject;
 import interface_adapter.PlayGameController;
 import interface_adapter.search.SearchViewModel;
-import use_case.note.search.SearchInteractor;
+import use_case.search.SearchInteractor;
 import view.MainFrame;
 
 public class GameApplication {
