+++ src/main/java/use_case/playgame/FetchPlayerStatisticsInputBoundary.java
@@ -1,4 +1,4 @@
-package use_case.note;
+package use_case.playgame;
 
 import java.util.List;
