+++ src/main/java/view/HomePanel.java
@@ -1,11 +1,13 @@
 package view;
 
+import app.Session;
+
 import javax.swing.*;
 import java.awt.*;
 
 public class HomePanel extends JPanel {
     public HomePanel(MainFrame frame) {
-        setLayout(new GridLayout(1, 3));
+        setLayout(new GridLayout(1, 4)); // Adjusted layout to fit an extra button
 
         // Play Game Button
         JButton playButton = new JButton("Play Game");
@@ -19,5 +21,16 @@ public class HomePanel extends JPanel {
 
         // Welcome Label
         add(new JLabel("Welcome to the Game App"));
+
+        // Logout Button
+        JButton logoutButton = new JButton("Logout");
+        logoutButton.addActionListener(e -> {
+            // Clear session data
+            Session.getInstance().clear();
+
+            // Navigate to LoginView
+            frame.switchTo("Login");
+        });
+        add(logoutButton);
     }
 }
