+++ src/main/java/view/MainFrame.java
@@ -1,10 +1,15 @@
 package view;
 
+import data_access.AccountDataAccessObject;
 import interface_adapter.PlayGameController;
-import use_case.note.search.SearchInteractor;
+import interface_adapter.account.AccountController;
+import interface_adapter.account.AccountPresenter;
+import use_case.account.AccountInteractor;
+import use_case.search.SearchInteractor;
 
 import javax.swing.*;
 import java.awt.*;
+import java.awt.event.ActionListener;
 
 public class MainFrame extends JFrame {
     private final CardLayout cardLayout = new CardLayout();
@@ -16,19 +21,41 @@ public class MainFrame extends JFrame {
         setDefaultCloseOperation(EXIT_ON_CLOSE);
 
         // Add panels to the CardLayout
-        mainPanel.add(new HomePanel(this), "Home");       // Home panel
+        mainPanel.add(new HomePanel(this), "Home");        // Home panel
         mainPanel.add(new PlayGamePanel(this, controller), "Play"); // PlayGame panel
         mainPanel.add(new SearchPanel(this, searchInteractor), "Search"); // Search panel
+
+        // Add LoginView
+        AccountController accountController = initializeAccountController();
+        ActionListener switchToSignUp = e -> switchTo("SignUp"); // Stub for switching to sign-up
+        LoginView loginView = new LoginView(accountController, switchToSignUp);
+        mainPanel.add(loginView, "Login");
 
         add(mainPanel);
     }
 
+    /**
+     * Switch to a specific panel by name.
+     */
     public void switchTo(String panelName) {
         cardLayout.show(mainPanel, panelName);
     }
 
+    /**
+     * Initialize the main frame with the Login panel visible.
+     */
     public void start() {
         setVisible(true);
-        switchTo("Home"); // Start with the Home panel
+        switchTo("Login"); // Start with the Login panel
+    }
+
+    /**
+     * Initializes the AccountController for managing login and account creation.
+     */
+    private AccountController initializeAccountController() {
+        AccountDataAccessObject accountDataAccess = new AccountDataAccessObject();
+        AccountPresenter accountPresenter = new AccountPresenter(this, new JFrame("Login")); // Pass MainFrame and login JFrame
+        AccountInteractor accountInteractor = new AccountInteractor(accountDataAccess, accountPresenter);
+        return new AccountController(accountInteractor);
     }
 }
