+++ src/main/java/view/SearchPanel.java
@@ -1,6 +1,6 @@
 package view;
 
-import use_case.note.search.SearchInteractor;
+import use_case.search.SearchInteractor;
 
 import javax.swing.*;
 import java.awt.*;
